+++ examples/bentoml_deployment/README.md
@@ -221,7 +221,7 @@ $ zenml model-deployer models describe cd38d6e6-467b-46e0-be13-3112c6e65d0e
 ┠────────────────────────┼────────────────────────────────────────────────────────────────────────────────┨
 ┃ PIPELINE_STEP_NAME     │ bentoml_model_deployer_step                                                    ┃
 ┠────────────────────────┼────────────────────────────────────────────────────────────────────────────────┨
-┃ PREDICITON_APIS_URLS   │ http://127.0.0.1:3001/predict_ndarray  http://127.0.0.1:3001/predict_image     ┃
+┃ PREDICTION_APIS_URLS   │ http://127.0.0.1:3001/predict_ndarray  http://127.0.0.1:3001/predict_image     ┃
 ┠────────────────────────┼────────────────────────────────────────────────────────────────────────────────┨
 ┃ PREDICTION_URL         │ http://127.0.0.1:3001/                                                         ┃
 ┠────────────────────────┼────────────────────────────────────────────────────────────────────────────────┨
