+++ examples/custom_code_deployment/run_kserve_pytorch.py
@@ -27,8 +27,8 @@ from kserve_pytorch.steps.inference_image_loader import (
     InferenceImageLoaderStepParameters,
     inference_image_loader,
 )
-from kserve_pytorch.steps.predection_service_loader import (
-    PredectionServiceLoaderStepParameters,
+from kserve_pytorch.steps.prediction_service_loader import (
+    PredictionServiceLoaderStepParameters,
     kserve_prediction_service_loader,
 )
 from kserve_pytorch.steps.predictor import kserve_predictor
@@ -158,7 +158,7 @@ def main(
                 ),
             ),
             prediction_service_loader=kserve_prediction_service_loader(
-                PredectionServiceLoaderStepParameters(
+                PredictionServiceLoaderStepParameters(
                     pipeline_name=deployment_pipeline_name,
                     step_name=step_name,
                     model_name=model_name,
