+++ examples/custom_code_deployment/run_kserve_tensorflow.py
@@ -29,8 +29,8 @@ from kserve_tensorflow.steps.inference_image_loader import (
     InferenceImageLoaderStepParameters,
     inference_image_loader,
 )
-from kserve_tensorflow.steps.predection_service_loader import (
-    PredectionServiceLoaderStepParameters,
+from kserve_tensorflow.steps.prediction_service_loader import (
+    PredictionServiceLoaderStepParameters,
     kserve_prediction_service_loader,
 )
 from kserve_tensorflow.steps.predictor import kserve_predictor
@@ -149,7 +149,7 @@ def main(
                 ),
             ),
             prediction_service_loader=kserve_prediction_service_loader(
-                PredectionServiceLoaderStepParameters(
+                PredictionServiceLoaderStepParameters(
                     pipeline_name=deployment_pipeline_name,
                     step_name=step_name,
                     model_name=model_name,
