+++ examples/custom_code_deployment/run_seldon_tensorflow.py
@@ -30,8 +30,8 @@ from seldon_tensorflow.steps.inference_image_loader import (
     InferenceImageLoaderStepParameters,
     inference_image_loader,
 )
-from seldon_tensorflow.steps.predection_service_loader import (
-    PredectionServiceLoaderStepParameters,
+from seldon_tensorflow.steps.prediction_service_loader import (
+    PredictionServiceLoaderStepParameters,
     seldon_prediction_service_loader,
 )
 from seldon_tensorflow.steps.predictor import seldon_predictor
@@ -149,7 +149,7 @@ def main(
                 ),
             ),
             prediction_service_loader=seldon_prediction_service_loader(
-                PredectionServiceLoaderStepParameters(
+                PredictionServiceLoaderStepParameters(
                     pipeline_name=deployment_pipeline_name,
                     step_name=step_name,
                     model_name=model_name,
