+++ examples/custom_code_deployment/seldon_tensorflow/steps/prediction_service_loader.py
@@ -13,13 +13,17 @@
 #  permissions and limitations under the License.
 from typing import cast
 
-from zenml.integrations.kserve.model_deployers import KServeModelDeployer
-from zenml.integrations.kserve.services import KServeDeploymentService
+from zenml.integrations.seldon.model_deployers.seldon_model_deployer import (
+    SeldonModelDeployer,
+)
+from zenml.integrations.seldon.services.seldon_deployment import (
+    SeldonDeploymentService,
+)
 from zenml.steps import BaseParameters, step
 
 
-class PredectionServiceLoaderStepParameters(BaseParameters):
-    """Prediction Service loader parameters.
+class PredictionServiceLoaderStepParameters(BaseParameters):
+    """Prediction Service loader configuration.
 
     Attrs:
         pipeline_name: name of the pipeline that deployed the model.
@@ -33,11 +37,11 @@ class PredectionServiceLoaderStepParameters(BaseParameters):
 
 
 @step(enable_cache=False)
-def kserve_prediction_service_loader(
-    params: PredectionServiceLoaderStepParameters,
-) -> KServeDeploymentService:
-    """Get the KServe prediction service started by the deployment pipeline."""
-    model_deployer = KServeModelDeployer.get_active_model_deployer()
+def seldon_prediction_service_loader(
+    params: PredictionServiceLoaderStepParameters,
+) -> SeldonDeploymentService:
+    """Get the Seldon Core prediction service started by the deployment pipeline."""
+    model_deployer = SeldonModelDeployer.get_active_model_deployer()
 
     services = model_deployer.find_model_server(
         pipeline_name=params.pipeline_name,
@@ -46,7 +50,7 @@ def kserve_prediction_service_loader(
     )
     if not services:
         raise RuntimeError(
-            f"No KServe prediction server deployed by the "
+            f"No Seldon Core prediction server deployed by the "
             f"'{params.step_name}' step in the '{params.pipeline_name}' "
             f"pipeline for the '{params.model_name}' model is currently "
             f"running."
@@ -54,10 +58,10 @@ def kserve_prediction_service_loader(
 
     if not services[0].is_running:
         raise RuntimeError(
-            f"The KServe prediction server last deployed by the "
+            f"The Seldon Core prediction server last deployed by the "
             f"'{params.step_name}' step in the '{params.pipeline_name}' "
             f"pipeline for the '{params.model_name}' model is not currently "
             f"running."
         )
 
-    return cast(KServeDeploymentService, services[0])
+    return cast(SeldonDeploymentService, services[0])
