+++ examples/kserve_deployment/run_pytorch.py
@@ -24,7 +24,7 @@ from steps.deployment_trigger import (
     deployment_trigger,
 )
 from steps.prediction_service_loader import (
-    PredectionServiceLoaderStepParameters,
+    PredictionServiceLoaderStepParameters,
     prediction_service_loader,
 )
 from steps.pytorch_steps import (
@@ -141,7 +141,7 @@ def main(
                 ),
             ),
             prediction_service_loader=prediction_service_loader(
-                PredectionServiceLoaderStepParameters(
+                PredictionServiceLoaderStepParameters(
                     pipeline_name=deployment_pipeline_name,
                     step_name=deployer_step_name,
                     model_name=model_name,
