+++ examples/kserve_deployment/steps/prediction_service_loader/prediction_service_loader.py
@@ -18,7 +18,7 @@ from zenml.integrations.kserve.services import KServeDeploymentService
 from zenml.steps import BaseParameters, step
 
 
-class PredectionServiceLoaderStepParameters(BaseParameters):
+class PredictionServiceLoaderStepParameters(BaseParameters):
     """KServe deployment loader parameters.
 
     Attributes:
@@ -36,7 +36,7 @@ class PredectionServiceLoaderStepParameters(BaseParameters):
 
 @step(enable_cache=False)
 def prediction_service_loader(
-    params: PredectionServiceLoaderStepParameters,
+    params: PredictionServiceLoaderStepParameters,
 ) -> KServeDeploymentService:
     """Get the prediction service started by the deployment pipeline."""
     model_deployer = KServeModelDeployer.get_active_model_deployer()
